Rename packages to conform to kotlin naming conventions
