+++ sample/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
             </intent-filter>
         </activity>
 
-        <service android:name="com.klaviyo.push_fcm.KlaviyoPushService" android:exported="false">
+        <service android:name="com.klaviyo.pushFcm.KlaviyoPushService" android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
