+++ sdk/analytics/src/test/java/com/klaviyo/analytics/KlaviyoTest.kt
@@ -11,8 +11,8 @@ import com.klaviyo.analytics.model.ProfileKey
 import com.klaviyo.analytics.networking.ApiClient
 import com.klaviyo.core.Registry
 import com.klaviyo.core.config.Config
-import com.klaviyo.core_shared_tests.BaseTest
-import com.klaviyo.core_shared_tests.StaticClock
+import com.klaviyo.fixtures.BaseTest
+import com.klaviyo.fixtures.StaticClock
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
