+++ sdk/analytics/src/test/java/com/klaviyo/analytics/model/ModelTests.kt
@@ -1,6 +1,6 @@
 package com.klaviyo.analytics.model
 
-import com.klaviyo.core_shared_tests.BaseTest
+import com.klaviyo.fixtures.BaseTest
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
