+++ sdk/core/build.gradle
@@ -1,3 +1,6 @@
+dependencies {
+    testImplementation project(':sdk:fixtures')
+}
 description = 'Core featureset of the Klaviyo SDK including SDK configuration, session tracking and analytics'
 evaluationDependsOn(':sdk')
 
@@ -14,9 +17,6 @@ android {
     }
 }
 
-dependencies {
-    testImplementation project(":sdk:core-shared-tests")
-}
 
 afterEvaluate {
     publishing {
