+++ sdk/core/src/test/java/com/klaviyo/core/model/SharedPreferencesDataStoreTest.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.content.SharedPreferences
 import com.klaviyo.core.Registry
 import com.klaviyo.core.model.SharedPreferencesDataStore.KLAVIYO_PREFS_NAME
-import com.klaviyo.core_shared_tests.BaseTest
+import com.klaviyo.fixtures.BaseTest
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.unmockkObject
