+++ sdk/push-fcm/build.gradle
@@ -18,8 +18,8 @@ dependencies {
     implementation project(':sdk:core')
     implementation project(':sdk:analytics')
     implementation "com.google.firebase:firebase-messaging:$ext.fcmVersion"
+    testImplementation project(':sdk:fixtures')
 
-    testImplementation project(":sdk:core-shared-tests")
 }
 
 afterEvaluate {
