+++ settings.gradle
@@ -1,4 +1,4 @@
 // sdk modules
 include ':sdk'
-include ':sdk:core', ':sdk:analytics', ':sdk:push-fcm', ':sdk:core-shared-tests'
+include ':sdk:core', ':sdk:analytics', ':sdk:push-fcm', ':sdk:fixtures'
 include ':sample'
